+++ src/Symfony/Component/Form/CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGELOG
    exception in 5.0.
  * Using names for buttons that do not contain only letters, digits, underscores, hyphens, and colons is deprecated and
    will lead to an exception in 5.0.
+ * added `html5` option to `NumberType` that allows to render `type="number"` input fields
  * deprecated using the `date_format`, `date_widget`, and `time_widget` options of the `DateTimeType` when the `widget`
    option is set to `single_text`
  * added `block_prefix` option to `BaseType`.